refactor: split into components and hooks
