+++ packages/web-components/lib/components/selfie/src/selfie-booth/components/index.ts
@@ -1,0 +1,3 @@
+export * from './CameraPreview';
+export * from './AlertDisplay';
+export * from './CaptureControls';
