+++ packages/web-components/lib/components/selfie/src/selfie-booth/hooks/index.ts
@@ -1,0 +1,2 @@
+export * from './useCamera';
+export * from './useFaceCapture';
