+++ packages/web-components/lib/components/selfie/src/selfie-booth/utils/index.ts
@@ -1,0 +1,4 @@
+export * from './faceDetection';
+export * from './canvas';
+export * from './imageCapture';
+export * from './alertMessages';
